Event/backend-django/requirements.txt
# Django framework
django==4.2.5
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt[redis]==5.3.1
django-extensions==3.2.3
whitenoise==6.5.0
django-taggit==4.0.0
markdown==3.4.3
django-filter==24.2
cerberus==1.3.5
argon2-cffi==23.1.0

# Documentation
drf-yasg==1.21.6
coreapi==2.3.3
mkdocs==1.5.0
sphinx==7.2.5

# Database
psycopg2==2.9.6

# Utils
mixer==7.2.2

# Images
pillow==10.3.0

# Linting
ruff==0.3.7
flake8==6.0.0

# Testing
pytest==7.4.0
pytest-django==4.5.2
factory-boy==3.2.1

# Redis support
redis==5.0.1
django-redis==5.3.0
